+++ src/gpu/VirtualFluids_GPU/GPU/KineticEnergyAnalyzer.cu
@@ -35,7 +35,7 @@ bool KineticEnergyAnalyzer::run(uint iter)
     if( iter % this->analyzeIter != 0 ) return false;
 
 	int lev = 0;
-	int size_Mat = this->para->getParD(lev)->numberOfNodes;
+	int size_Mat = (int)this->para->getParD(lev)->numberOfNodes;
 
     thrust::device_vector<real> kineticEnergy(size_Mat, c0o1);
     thrust::device_vector<uint> isFluid      (size_Mat, 0);
