+++ src/gpu/VirtualFluids_GPU/Kernel/Kernels/BasicKernels/FluidFlow/Compressible/CumulantK15Sponge/CumulantK15SpongeComp.cu
@@ -10,7 +10,7 @@ std::shared_ptr<CumulantK15SpongeComp> CumulantK15SpongeComp::getNewInstance(std
 
 void CumulantK15SpongeComp::run()
 {
-	int size_Mat = para->getParD(level)->numberOfNodes;
+	int size_Mat = (int)para->getParD(level)->numberOfNodes;
 	int numberOfThreads = para->getParD(level)->numberofthreads;
 
 	int Grid = (size_Mat / numberOfThreads) + 1;
@@ -28,17 +28,18 @@ void CumulantK15SpongeComp::run()
 	dim3 grid(Grid1, Grid2);
 	dim3 threads(numberOfThreads, 1, 1);
 
-	LB_Kernel_CumulantK15SpongeComp <<< grid, threads >>>(	para->getParD(level)->omega,
-															para->getParD(level)->typeOfGridNode,
-															para->getParD(level)->neighborX,
-															para->getParD(level)->neighborY,
-															para->getParD(level)->neighborZ,
-															para->getParD(level)->coordinateX,
-															para->getParD(level)->coordinateY,
-															para->getParD(level)->coordinateZ,
-															para->getParD(level)->distributions.f[0],
-															para->getParD(level)->numberOfNodes,
-															para->getParD(level)->isEvenTimestep);
+	LB_Kernel_CumulantK15SpongeComp <<< grid, threads >>>(
+		para->getParD(level)->omega,
+		para->getParD(level)->typeOfGridNode,
+		para->getParD(level)->neighborX,
+		para->getParD(level)->neighborY,
+		para->getParD(level)->neighborZ,
+		para->getParD(level)->coordinateX,
+		para->getParD(level)->coordinateY,
+		para->getParD(level)->coordinateZ,
+		para->getParD(level)->distributions.f[0],
+		para->getParD(level)->numberOfNodes,
+		para->getParD(level)->isEvenTimestep);
 	getLastCudaError("LB_Kernel_CumulantK15SpongeComp execution failed");
 }
