+++ src/gpu/VirtualFluids_GPU/Kernel/Kernels/WaleKernels/FluidFlow/Compressible/CumulantK15/WaleCumulantK15Comp.cu
@@ -10,7 +10,7 @@ std::shared_ptr<WaleCumulantK15Comp> WaleCumulantK15Comp::getNewInstance(std::sh
 
 void WaleCumulantK15Comp::run()
 {
-	int size_Mat = para->getParD(level)->numberOfNodes;
+	int size_Mat = (int)para->getParD(level)->numberOfNodes;
 	int numberOfThreads = para->getParD(level)->numberofthreads;
 
 	int Grid = (size_Mat / numberOfThreads) + 1;
@@ -28,22 +28,23 @@ void WaleCumulantK15Comp::run()
 	dim3 grid(Grid1, Grid2, 1);
 	dim3 threads(numberOfThreads, 1, 1);
 
-	LB_Kernel_WaleCumulantK15Comp << < grid, threads >> >(	para->getParD(level)->omega,
-																para->getParD(level)->typeOfGridNode,
-																para->getParD(level)->neighborX,
-																para->getParD(level)->neighborY,
-																para->getParD(level)->neighborZ,
-																para->getParD(level)->neighborInverse,
-																para->getParD(level)->velocityX,
-																para->getParD(level)->velocityY,
-																para->getParD(level)->velocityZ,
-																para->getParD(level)->distributions.f[0],
-																para->getParD(level)->turbViscosity,
-																para->getParD(level)->numberOfNodes,
-																level,
-																para->getTimestepOfCoarseLevel(),
-																para->getForcesDev(),
-																para->getParD(level)->isEvenTimestep);
+	LB_Kernel_WaleCumulantK15Comp <<< grid, threads >>>(
+		para->getParD(level)->omega,
+		para->getParD(level)->typeOfGridNode,
+		para->getParD(level)->neighborX,
+		para->getParD(level)->neighborY,
+		para->getParD(level)->neighborZ,
+		para->getParD(level)->neighborInverse,
+		para->getParD(level)->velocityX,
+		para->getParD(level)->velocityY,
+		para->getParD(level)->velocityZ,
+		para->getParD(level)->distributions.f[0],
+		para->getParD(level)->turbViscosity,
+		para->getParD(level)->numberOfNodes,
+		level,
+		para->getTimestepOfCoarseLevel(),
+		para->getForcesDev(),
+		para->getParD(level)->isEvenTimestep);
 	getLastCudaError("LB_Kernel_WaleCumulantK15Comp execution failed");
 }
